pyproject.toml
[tool.poetry]
name = "contact-scraper"
version = "0.1.0"
description = ""
authors = ["John Finberg <johnbfinberg@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
click = "^8.1.8"
requests = "^2.32.3"
html2text = "^2024.2.26"
openai = "^1.64.0"
beautifulsoup4 = "^4.13.3"
pydantic = "^2.10.6"
python-dotenv = "^1.0.1"

[tool.poetry.scripts]
contact-scraper = "contact_scraper.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
